requirements.txt
art==5.4
colorama==0.4.4
Pillow==9.0.0
termcolor==1.1.0
